# openapi.yaml
openapi: 3.0.3
info:
  title: ML Virtual Lab API
  version: "0.2.0"
  description: |
    Backend for the ML Virtual Lab. Three roles: **student**, **instructor**,
    **admin** — see `CLAUDE.md` for the full onboarding/permission model.

    ### Status
    Paths marked `x-status: implemented` exist in `app/routers/` today.
    Paths marked `x-status: planned` are the agreed target design (matches
    the four starter issues + this onboarding issue) but are not yet coded —
    treat their shape as a contract to build against, not as fact about
    what's live.

    ### Auth
    All endpoints except `/health`, `POST /auth/login` and
    `POST /batches/{batch_id}/students`'s *targets* (the students being
    created, not the caller) require a Bearer JWT. A student whose account
    still has its emailed default password (`must_change_password: true`)
    can only call `/auth/me` and `/auth/change-password` — every other
    endpoint returns 403 until it's changed.
servers:
  - url: http://localhost:8000
    description: Local dev
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Log in with email + password
      security: []
      x-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
            examples:
              default: { value: { email: student1@college.edu, password: "TRbB5JGVwS" } }
      responses:
        "200":
          description: Login succeeded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "401":
          description: Wrong email or password (same message for both, deliberately)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { detail: "Invalid credentials" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /auth/me:
    get:
      tags: [auth]
      summary: The logged-in user's own profile
      x-status: implemented
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/change-password:
    post:
      tags: [auth]
      summary: Change your own password (required before anything else, for a new student account)
      x-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      responses:
        "200":
          description: Password changed; must_change_password is now false
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          description: current_password is wrong, or not logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: new_password too short, or same as current_password
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ---------------- BATCHES ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Missing, invalid or expired token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { detail: "Not authenticated" }
...
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        detail: { type: string, example: "Invalid credentials" }
      required: [detail]
    UserRole:
      type: string
      enum: [student, instructor, admin]
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        email: { type: string, format: email }
        roll_number: { type: string, nullable: true }
        role: { $ref: "#/components/schemas/UserRole" }
        is_active: { type: boolean }
        must_change_password:
          type: boolean
          description: True until the student changes their emailed default password.
      required: [id, full_name, email, role, is_active, must_change_password]
    LoginRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string }
      required: [email, password]
    TokenResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: bearer }
        must_change_password: { type: boolean }
      required: [access_token, token_type, must_change_password]
    ChangePasswordRequest:
      type: object
      properties:
        current_password: { type: string }
        new_password: { type: string, minLength: 8 }
      required: [current_password, new_password]
...